% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{as.data.frame.vcs_dictionary}
\alias{as.data.frame.vcs_dictionary}
\title{Convert a dictionary to a data frame}
\usage{
\method{as.data.frame}{vcs_dictionary}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=vcs_dictionary]{vcs_dictionary()}}.}

\item{...}{Unused.}
}
\value{
A tibble with columns \code{concept}, \code{source}, \code{level} and \code{required}.
}
\description{
Convert a dictionary to a data frame
}
\examples{
as.data.frame(vcs_dictionary(child_id = "cid", psu = "clust"))
}
