% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{as_survey_design}
\alias{as_survey_design}
\title{Extract the underlying survey design object}
\usage{
as_survey_design(x)
}
\arguments{
\item{x}{A \code{\link[=vcs_design]{vcs_design()}}.}
}
\value{
A \code{survey.design} object.
}
\description{
Use this to reach any function in the \code{survey} package that vaxsurvR does not
wrap.
}
\examples{
des <- vcs_design(derive_vaccination_status(vcs_example))
survey::svymean(~cov_BCG, as_survey_design(des), na.rm = TRUE)
}
