% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_table.R
\name{bar_table_size}
\alias{bar_table_size}
\title{Suggested figure size for a bar table, in inches}
\usage{
bar_table_size(p)
}
\arguments{
\item{p}{A plot from \code{\link[=plot_bar_table]{plot_bar_table()}}.}
}
\value{
A named numeric vector \code{width}, \code{height}.
}
\description{
Suggested figure size for a bar table, in inches
}
