% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{c.vcs_validation}
\alias{c.vcs_validation}
\title{Combine validation results}
\usage{
\method{c}{vcs_validation}(...)
}
\arguments{
\item{...}{\link[=new_vcs_validation]{vcs_validation} objects.}
}
\value{
A single \code{vcs_validation}.
}
\description{
Combine validation results
}
\examples{
d <- vcs_example
c(check_unique_ids(d), check_duplicates(d))
}
