% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{card_completeness}
\alias{card_completeness}
\title{Card availability and completeness}
\usage{
card_completeness(x, by = NULL)
}
\arguments{
\item{x}{A \link[=new_vcs_data]{vcs_data} object.}

\item{by}{Grouping variables.}
}
\value{
A tibble with \code{n_children}, \code{n_card_seen}, \code{prop_card_seen},
\code{n_card_missing} and \code{prop_card_missing} per group.
}
\description{
Card availability and completeness
}
\examples{
card_completeness(vcs_example)
card_completeness(vcs_example, by = ~stratum)
}
