% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_qa.R
\name{check_dose_intervals}
\alias{check_dose_intervals}
\title{Check minimum intervals between consecutive doses}
\usage{
check_dose_intervals(x, schedule = NULL, tolerance_days = 0)
}
\arguments{
\item{x}{A \link[=new_vcs_data]{vcs_data} object.}

\item{schedule}{A \code{\link[=vcs_schedule]{vcs_schedule()}}. Defaults to the schedule stored in \code{x}.}

\item{tolerance_days}{Slack allowed below the schedule minimum interval.}
}
\value{
A \link[=new_vcs_validation]{vcs_validation} object.
}
\description{
Check minimum intervals between consecutive doses
}
\examples{
check_dose_intervals(vcs_example)
}
\seealso{
Other vaccination checks:
\code{\link[=check_age_at_vaccination]{check_age_at_vaccination()}},
\code{\link[=check_card_transcription]{check_card_transcription()}},
\code{\link[=check_duplicate_doses]{check_duplicate_doses()}},
\code{\link[=check_future_dates]{check_future_dates()}},
\code{\link[=check_prebirth_vaccination]{check_prebirth_vaccination()}},
\code{\link[=check_vaccine_dates]{check_vaccine_dates()}},
\code{\link[=check_vaccine_sequence]{check_vaccine_sequence()}},
\code{\link[=flag_vaccine_inconsistencies]{flag_vaccine_inconsistencies()}}
}
\concept{vaccination checks}
