% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_qa.R
\name{check_duplicate_doses}
\alias{check_duplicate_doses}
\title{Check for duplicated doses on the same day}
\usage{
check_duplicate_doses(x)
}
\arguments{
\item{x}{A \link[=new_vcs_data]{vcs_data} object.}
}
\value{
A \link[=new_vcs_validation]{vcs_validation} object.
}
\description{
Two doses of the \emph{same antigen} recorded on the same date are almost always
a transcription error: the same line of the card read twice.
}
\examples{
check_duplicate_doses(vcs_example)
}
\seealso{
Other vaccination checks:
\code{\link[=check_age_at_vaccination]{check_age_at_vaccination()}},
\code{\link[=check_card_transcription]{check_card_transcription()}},
\code{\link[=check_dose_intervals]{check_dose_intervals()}},
\code{\link[=check_future_dates]{check_future_dates()}},
\code{\link[=check_prebirth_vaccination]{check_prebirth_vaccination()}},
\code{\link[=check_vaccine_dates]{check_vaccine_dates()}},
\code{\link[=check_vaccine_sequence]{check_vaccine_sequence()}},
\code{\link[=flag_vaccine_inconsistencies]{flag_vaccine_inconsistencies()}}
}
\concept{vaccination checks}
