% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_qa.R
\name{check_future_dates}
\alias{check_future_dates}
\title{Check for vaccination dates in the future}
\usage{
check_future_dates(x, reference = NULL)
}
\arguments{
\item{x}{A \link[=new_vcs_data]{vcs_data} object.}

\item{reference}{Date after which a vaccination date is impossible. Defaults
to each child's interview date where available, otherwise today.}
}
\value{
A \link[=new_vcs_validation]{vcs_validation} object.
}
\description{
Check for vaccination dates in the future
}
\examples{
check_future_dates(vcs_example)
}
\seealso{
Other vaccination checks:
\code{\link[=check_age_at_vaccination]{check_age_at_vaccination()}},
\code{\link[=check_card_transcription]{check_card_transcription()}},
\code{\link[=check_dose_intervals]{check_dose_intervals()}},
\code{\link[=check_duplicate_doses]{check_duplicate_doses()}},
\code{\link[=check_prebirth_vaccination]{check_prebirth_vaccination()}},
\code{\link[=check_vaccine_dates]{check_vaccine_dates()}},
\code{\link[=check_vaccine_sequence]{check_vaccine_sequence()}},
\code{\link[=flag_vaccine_inconsistencies]{flag_vaccine_inconsistencies()}}
}
\concept{vaccination checks}
