% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_structural.R
\name{check_required_variables}
\alias{check_required_variables}
\title{Check that required concepts were mapped and are populated}
\usage{
check_required_variables(x, required = NULL)
}
\arguments{
\item{x}{A \link[=new_vcs_data]{vcs_data} object.}

\item{required}{Concepts that must be present. Defaults to the concepts
marked required in \code{\link[=vcs_concepts]{vcs_concepts()}}.}
}
\value{
A \link[=new_vcs_validation]{vcs_validation} object.
}
\description{
Check that required concepts were mapped and are populated
}
\examples{
check_required_variables(vcs_example)
}
\seealso{
Other structural checks:
\code{\link[=check_child_eligibility]{check_child_eligibility()}},
\code{\link[=check_duplicates]{check_duplicates()}},
\code{\link[=check_household_structure]{check_household_structure()}},
\code{\link[=check_psu_structure]{check_psu_structure()}},
\code{\link[=check_segment_structure]{check_segment_structure()}},
\code{\link[=check_unique_ids]{check_unique_ids()}}
}
\concept{structural checks}
