% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_structural.R
\name{check_segment_structure}
\alias{check_segment_structure}
\title{Check the segment structure within PSUs}
\usage{
check_segment_structure(x, max_segments = 10L)
}
\arguments{
\item{x}{A \link[=new_vcs_data]{vcs_data} object.}

\item{max_segments}{Number of segments per PSU above which the PSU is
flagged.}
}
\value{
A \link[=new_vcs_validation]{vcs_validation} object.
}
\description{
Check the segment structure within PSUs
}
\examples{
check_segment_structure(vcs_example)
}
\seealso{
Other structural checks:
\code{\link[=check_child_eligibility]{check_child_eligibility()}},
\code{\link[=check_duplicates]{check_duplicates()}},
\code{\link[=check_household_structure]{check_household_structure()}},
\code{\link[=check_psu_structure]{check_psu_structure()}},
\code{\link[=check_required_variables]{check_required_variables()}},
\code{\link[=check_unique_ids]{check_unique_ids()}}
}
\concept{structural checks}
