% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcs_data.R
\name{derivation_rules}
\alias{derivation_rules}
\title{List the derivation rules recorded in a table}
\usage{
derivation_rules(x)
}
\arguments{
\item{x}{A data frame, or a \link[=new_vcs_data]{vcs_data} object.}
}
\value{
A tibble with columns \code{variable} and \code{rule}.
}
\description{
List the derivation rules recorded in a table
}
\examples{
derivation_rules(vcs_example)
}
