% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_full_coverage}
\alias{estimate_full_coverage}
\title{Estimate the proportion fully vaccinated}
\usage{
estimate_full_coverage(design, variable = "fully_vaccinated", by = NULL, ...)
}
\arguments{
\item{design}{A \code{\link[=vcs_design]{vcs_design()}} or \link[=new_vcs_data]{vcs_data} object.}

\item{variable}{Name of the fully-vaccinated indicator, as created by
\code{\link[=derive_fully_vaccinated]{derive_fully_vaccinated()}}.}

\item{by}{Domains.}

\item{...}{Passed to the underlying estimator (\code{level}, \code{deff}, \code{na.rm}).}
}
\value{
A \code{vcs_estimate} tibble.
}
\description{
Estimate the proportion fully vaccinated
}
\examples{
d <- derive_fully_vaccinated(vcs_example)
estimate_full_coverage(vcs_design(d))
}
