% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_timely_coverage}
\alias{estimate_timely_coverage}
\title{Estimate timely coverage}
\usage{
estimate_timely_coverage(
  design,
  vaccines = NULL,
  by = NULL,
  prefix = "timely_",
  ...
)
}
\arguments{
\item{design}{A \code{\link[=vcs_design]{vcs_design()}} or \link[=new_vcs_data]{vcs_data} object.}

\item{vaccines}{Vaccine names.}

\item{by}{Domains.}

\item{prefix}{Prefix of the timeliness columns, as created by
\code{\link[=derive_timeliness]{derive_timeliness()}}.}

\item{...}{Passed to the underlying estimator.}
}
\value{
A \code{vcs_estimate} tibble.
}
\description{
Estimate timely coverage
}
\examples{
d <- derive_timeliness(vcs_example)
estimate_timely_coverage(vcs_design(d), vaccines = c("BCG", "PENTA1"))
}
