% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_zero_dose}
\alias{estimate_zero_dose}
\title{Estimate the zero-dose proportion}
\usage{
estimate_zero_dose(design, variable = "zero_dose", by = NULL, ...)
}
\arguments{
\item{design}{A \code{\link[=vcs_design]{vcs_design()}} or \link[=new_vcs_data]{vcs_data} object.}

\item{variable}{Name of the zero-dose indicator, as created by
\code{\link[=derive_zero_dose]{derive_zero_dose()}}.}

\item{by}{Domains.}

\item{...}{Passed to the underlying estimator.}
}
\value{
A \code{vcs_estimate} tibble.
}
\description{
Estimate the zero-dose proportion
}
\examples{
d <- derive_zero_dose(vcs_example)
estimate_zero_dose(vcs_design(d), by = ~stratum)
}
