% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour.R
\name{ft_reasons_table}
\alias{ft_reasons_table}
\title{Reasons-not-vaccinated table as a flextable}
\usage{
ft_reasons_table(
  x,
  palette = vcs_palette(),
  font_size = 8,
  total_width = getOption("vaxsurvR.table_width", 6.7)
)
}
\arguments{
\item{x}{Output of \code{\link[=reasons_by_stratum]{reasons_by_stratum()}}.}

\item{palette}{A \code{\link[=vcs_palette]{vcs_palette()}}.}

\item{font_size}{Body font size in points.}

\item{total_width}{Table width in inches.}
}
\value{
A \code{flextable}.
}
\description{
The Word/HTML counterpart of \code{\link[=plot_reasons_table]{plot_reasons_table()}}.
}
