% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{has_issues}
\alias{has_issues}
\title{Does a validation result contain blocking issues?}
\usage{
has_issues(x, severity = "ERROR")
}
\arguments{
\item{x}{A \link[=new_vcs_validation]{vcs_validation} object.}

\item{severity}{Minimum severity that counts as blocking.}
}
\value{
A logical scalar.
}
\description{
Does a validation result contain blocking issues?
}
\examples{
has_issues(validate_vcs(vcs_example), "CRITICAL")
}
