% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{is_vcs_audit}
\alias{is_vcs_audit}
\title{Test whether an object is a vcs_audit}
\usage{
is_vcs_audit(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_audit
}
\examples{
is_vcs_audit(vcs_log())
}
