% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcs_data.R
\name{is_vcs_data}
\alias{is_vcs_data}
\title{Test whether an object is a vcs_data}
\usage{
is_vcs_data(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_data
}
\examples{
is_vcs_data(vcs_example)
}
