% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{is_vcs_design}
\alias{is_vcs_design}
\title{Test whether an object is a vcs_design}
\usage{
is_vcs_design(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_design
}
\examples{
is_vcs_design(vcs_design(derive_vaccination_status(vcs_example)))
}
