% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{is_vcs_dictionary}
\alias{is_vcs_dictionary}
\title{Test whether an object is a vcs_dictionary}
\usage{
is_vcs_dictionary(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_dictionary
}
\examples{
is_vcs_dictionary(vcs_dictionary(child_id = "id"))
is_vcs_dictionary(list())
}
