% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{is_vcs_estimate}
\alias{is_vcs_estimate}
\title{Test whether an object is a vcs_estimate}
\usage{
is_vcs_estimate(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_estimate
}
\examples{
is_vcs_estimate(estimate_coverage(
  vcs_design(derive_vaccination_status(vcs_example)), vaccines = "BCG"
))
}
