% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{is_vcs_schedule}
\alias{is_vcs_schedule}
\title{Test whether an object is a vcs_schedule}
\usage{
is_vcs_schedule(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_schedule
}
\examples{
is_vcs_schedule(vcs_schedule_who())
}
