% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_vcs_validation}
\alias{is_vcs_validation}
\title{Test whether an object is a vcs_validation}
\usage{
is_vcs_validation(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a vcs_validation
}
\examples{
is_vcs_validation(new_vcs_validation())
}
