% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_kc9.R
\name{kc9_epi_doses}
\alias{kc9_epi_doses}
\title{Names of the EPI doses in the KC v9 schedule}
\usage{
kc9_epi_doses(malaria = TRUE)
}
\arguments{
\item{malaria}{Include the four malaria doses.}
}
\value{
A character vector in schedule order.
}
\description{
Names of the EPI doses in the KC v9 schedule
}
\examples{
kc9_epi_doses()
}
\seealso{
Other KC v9 configuration:
\code{\link[=attach_child_variables]{attach_child_variables()}},
\code{\link[=kc9_besd_map]{kc9_besd_map()}},
\code{\link[=kc9_caregiver_extras]{kc9_caregiver_extras()}},
\code{\link[=kc9_caregiver_to_child]{kc9_caregiver_to_child()}},
\code{\link[=kc9_child_extras]{kc9_child_extras()}},
\code{\link[=kc9_dictionary]{kc9_dictionary()}},
\code{\link[=kc9_labels]{kc9_labels()}},
\code{\link[=kc9_link_roster]{kc9_link_roster()}},
\code{\link[=kc9_prepare]{kc9_prepare()}},
\code{\link[=kc9_recall_map]{kc9_recall_map()}},
\code{\link[=kc9_schedule]{kc9_schedule()}},
\code{\link[=kc9_strata]{kc9_strata()}}
}
\concept{KC v9 configuration}
